+++ src/E5R.Architecture.Data.EntityFrameworkCore/RideStorage.cs
@@ -47,6 +47,9 @@ namespace E5R.Architecture.Data.EntityFrameworkCore
         public int Count(IDataFilter<TDataModel> filter) =>
             QuerySearch(_query, filter, null).Count();
 
+        public TDataModel GetFirst(IDataFilter<TDataModel> filter, IDataIncludes includes = null) =>
+            QuerySearch(_query, filter, includes).FirstOrDefault();
+
         public IEnumerable<TDataModel> GetAll(IDataIncludes includes) =>
             TryApplyIncludes(_query, includes).ToList();
 
@@ -116,6 +119,15 @@ namespace E5R.Architecture.Data.EntityFrameworkCore
         {
             throw new DataLayerException(
                 $"{this.GetType().Name} not implement {nameof(Find)}(identifiers)!");
+        }
+
+        public TSelect GetFirst<TSelect>(IDataFilter<TDataModel> filter,
+            IDataProjection<TDataModel, TSelect> projection)
+        {
+            Checker.NotNullArgument(projection, nameof(projection));
+            Checker.NotNullObject(projection.Select, () => projection.Select);
+
+            return QuerySearch(_query, filter, projection).Select(projection.Select).FirstOrDefault();
         }
 
         public IEnumerable<TSelect> GetAll<TSelect>(IDataProjection<TDataModel, TSelect> projection)