+++ src/E5R.Architecture.Data/Abstractions/IAcquirableStorageWithSelector.cs
@@ -19,6 +19,15 @@ namespace E5R.Architecture.Data.Abstractions
         where TDataModel : IIdentifiable
     {
         /// <summary>
+        /// Get the first item stored according to filter entered
+        /// </summary>
+        /// <param name="filter">Data filter</param>
+        /// <param name="projection">Data projection (Include and Select) for data model</param>
+        /// <returns>Instance of <see cref="IDataProjection{TDataModel, TGroup}.Select"/>, or null when not found</returns>
+        TSelect GetFirst<TSelect>(IDataFilter<TDataModel> filter,
+            IDataProjection<TDataModel, TSelect> projection);
+
+        /// <summary>
         /// Get all stored items
         /// </summary>
         /// <param name="projection">Data projection (Include and Select) for data model</param>